refactor(engine): split walk helpers out of mod.rs
